minor
Removed SOCK=0 condition to show upgraded strength next to upgrade recipe for Griswold's Heart, which has natural sockets. The rule is commented out, so it doesn't currently matter for the Feather filter. This is just an update to reduce confusion for anyone borrowing/modifying the code to make use of that rule.